requirements.txt
aiogram==2.25.1
beautifulsoup4==4.12.0
DateTime==5.1
Flask==3.0.0
google-api-python-client
httplib2
load-dotenv==0.1.0
oauth2client
openpyxl
requests==2.25.1
statistics==1.0.3.5
Telethon==1.28.5
pandas
pillow
selenium
numpy
